Addressed warnings and improved code quality with better null handling and edge case coverage

+++ Interfaces/IProductRepository.cs
@@ -1,6 +1,6 @@
 public interface IProductRepository
 {
-    double GetPrice(string productName);
+    decimal GetPrice(string productName);
 
     bool Exists(string productName);
 }
